moved terminal hooks into terminal domain
