+++ src/HiddenPages/TerminalEditor/DirectoryEditor.tsx
@@ -4,10 +4,10 @@ import './directory-editor.css';
 import { Button, ButtonTheme } from "../../SharedComponents/Button/Button";
 import { PlusIcon } from "../../SharedComponents/Icons/PlusIcon";
 import { TextInput } from "../../SharedComponents/TextInput/TextInput";
-import { Directory } from "./hooks/directories/useDirectories";
+import { Directory } from "../../Terminal/hooks/directories/useDirectories";
 import { Card } from "../../SharedComponents/Card/Card";
 import { ToastMessageContext } from "../../SharedComponents/Toast/ToastMessageContext";
-import { File, useFiles } from "./hooks/files/useFiles";
+import { File, useFiles } from "../../Terminal/hooks/files/useFiles";
 import { Loader } from "../../SharedComponents/Loader/Loader";
 import { FileEditorModal } from "./FileEditorModal";
 import { TrashIcon } from "../../SharedComponents/Icons/TrashIcon";
