+++ src/Terminal/hooks/directories/useDeleteDirectory.ts
