+++ src/Terminal/hooks/directories/useDirectories.ts
