+++ src/Terminal/hooks/files/useDeleteFile.ts
