+++ src/Terminal/hooks/server/useDeleteServer.ts
